+++ dir.c
@@ -395,7 +395,7 @@ Return Value:
         default:
 
             PutStdErr(MSG_PARAMETER_FORMAT_NOT_CORRECT, ONEARG,
-                      argstr1(Fmt14,(unsigned long)((int)(&(pszTok[irgch])))));
+                      argstr1(Fmt14,((&(pszTok[irgch])))));
             return( FAILURE );
 
         } // switch
@@ -520,7 +520,7 @@ Return Value:
         case MINUS:
             if (fOff) {
                 PutStdErr(MSG_PARAMETER_FORMAT_NOT_CORRECT, ONEARG,
-                          argstr1(Fmt14,(unsigned long)((int)(&(pszTok[irgch])))));
+                          argstr1(Fmt14,((&(pszTok[irgch])))));
                 return( FAILURE );
             }
 
@@ -530,7 +530,7 @@ Return Value:
         default:
 
             PutStdErr(MSG_PARAMETER_FORMAT_NOT_CORRECT, ONEARG,
-                      argstr1(Fmt14,(unsigned long)((int)(&(pszTok[irgch])))));
+                      argstr1(Fmt14,((&(pszTok[irgch])))));
             return( FAILURE );
 
         } // switch
@@ -617,7 +617,7 @@ Return Value:
             if (pszTok[irgch+1] == MINUS) {
 
                 PutStdErr(MSG_PARAMETER_FORMAT_NOT_CORRECT, ONEARG,
-                argstr1(Fmt14,(unsigned long)((int)(&(pszTok[irgch])))));
+                argstr1(Fmt14,((&(pszTok[irgch])))));
                 return( FAILURE );
 
             }
@@ -629,7 +629,7 @@ Return Value:
         default:
 
             PutStdErr(MSG_PARAMETER_FORMAT_NOT_CORRECT, ONEARG,
-            argstr1(Fmt14,(unsigned long)((int)(&(pszTok[irgch])))));
+            argstr1(Fmt14,((&(pszTok[irgch])))));
             return( FAILURE );
 
         } // switch
